pyproject.toml
[project]
name = "jin-yong-mud"
version = "0.1.0"
description = "Jin Yong Wuxia realtime MUD MVP with Evennia adapter"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
  "types-PyYAML>=6.0.12.20240808",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
strict = true

[tool.pytest.ini_options]
addopts = "-q"
